requirements.txt
fastapi
uvicorn
pydantic
openai
jinja2
pytz
google-genai
requests
langgraph
langchain-core
